% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correctedBed2AllInfo.R
\name{correctedBed2AllInfo}
\alias{correctedBed2AllInfo}
\alias{bed2AllInfo}
\title{Make a "corrected" version of the All-Info file
after using IsoQuant to identify incorrect splice sites}
\usage{
correctedBed2AllInfo(bedPath='Sample_sorted.corrected_reads.bed',
allInfo = "LongReadInfo/AllInfo_IncompleteReads_corrected.gz",
outFile = "LongReadInfo/AllInfo_IncompleteReads_corrected.gz")
}
\arguments{
\item{bedPath}{Path to corrected bed file. Currently does not
support bed.gz}

\item{Path}{to the output.gz file. Defaults to the LongReadInfo folder and
generates a file named "AllInfo_IncompleteReads_corrected.gz".}
}
\value{
A file in the AllInfo_Incomplete format with column 6 being the
corrected intron chain, column 9 being the corrected exon chain, and
additional columns 10 and 11 being the original intron and exon chains
respectively
}
\description{
This function takes in a corrected bed file
that is generated after running IsoQuant, as well as the AllInfo
file that is the output of scisorseqr. It produces a new AllInfo
file with extra columns containing the original and corrected exon
and intron chains. Best use case is for Oxford Nanopore data which
can have errors in splice-site assignment
}
\seealso{
\code{\link{InfoPerLongRead}}

\link[IsoQuant]{https://github.com/ablab/IsoQuant}
}
