% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExonQuant.R
\name{ExonQuant}
\alias{ExonQuant}
\title{Count exon inclusion and exclusion according to grouping factor}
\usage{
ExonQuant(
  allInfoFile = "LongReadInfo/AllInfo",
 
    exon.gff = "LRProcessingOutput/mapping.bestperRead.RNAdirection.withConsensIntrons.gff.gz",
  groupingFactor = "Celltype",
  threshold = 10
)
}
\arguments{
\item{allInfoFile}{file containing cell barcode and celltype information
per read. Defaults to output of the InfoPerLongRead function}

\item{exon.gff}{gzipped file containing exon mapping information.
Defaults to output from MapAndFilter}

\item{groupingFactor}{"Celltype" or "Barcode" to group reads by for
counting inclusion levels. Defaults to celltype}

\item{threshold}{minimum number of reads per grouping factor in order
to consider that exon to be sufficiently expressed. Defaults to 10}
}
\value{
ExonQuantOutput/InclusionExclusionCounts.tsv
}
\description{
This function groups reads by either the celltype
or cell-barcode to count inclusion and exclusion of exons. Inputs
to this function are based on previous preprocessing steps
}
\seealso{
\code{\link{MapAndFilter}} \code{\link{InfoPerLongRead}}
}
