% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FilterBCoutput.R
\name{FilterBCOutput}
\alias{FilterBCOutput}
\title{Filter unbarcoded reads}
\usage{
FilterBCOutput(outputFolder, raw_Output, concatenate = TRUE)
}
\arguments{
\item{outputFolder}{automatically sourced from GetBarcodes script}

\item{raw_Output}{unfiltered output of GetBarcodes script}

\item{concatenate}{Combines output of all fastq.gz files if they belong
to the same sample. Defaults to TRUE}
}
\description{
This function filters out all the reads that don't have a barcode
and outputs them into a separate folder.
Option to concatenate output from several files if they belong to
the same sample.
}
