% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetBarcodes.R
\name{GetBarcodes}
\alias{GetBarcodes}
\title{Get single-cell barcodes from long read files}
\usage{
GetBarcodes("FastqFolder","Barcode-Clust","~/MyDir/",6,concatenate = FALSE)
}
\arguments{
\item{fqFolder}{Path to a folder containing fastq.gz files}

\item{BCClustAssignFile}{tab separated file containing barcode in column 1
and cluster in column 2}

\item{outputFolder}{OPTIONAL Path to working directory. Defaults to current}

\item{numProcesses}{Number of chunks to split processing into. Defaults to 10}

\item{concatenate}{OPTIONAL Concatenate output from all files in the folder}
}
\value{
OutputRaw folder containing a .csv file and summary stats for each
input fastq.gz file

OutputFiltered Filtered file with one line for each read containing
a barcode. Option to concatenate into one file per input folder
}
\description{
This is an R-wrapper for a python function. It parallelizes the reading
and processing of fastq.gz files, and uses a sliding window approach to
identify cell barcodes, assign cluster labels and output useful statistics
in one .csv file and one _summary file per fastq.gz input file
}
