% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GO.Reference.R
\name{GO.Reference}
\alias{GO.Reference}
\title{Conducts gene set enrichment analysis of significantly spliced genes}
\usage{
GO.Reference('./','TreeTraversal_Iso/')
}
\arguments{
\item{working_directory}{where the OutputGO folder should be created}

\item{comparisonsDir}{location of TreeTraversal folder containing output
files from differential isoform expression analysis}
}
\value{
OutputGO containing folders for Molecular Function, Biological
Process, and Cellular Component

.Robj files for MF,BP, and Cc
}
\description{
An easy way to look for enriched genes in the differential
splicing analysis. Uses the ClusterProfiler package to compare geneset
against the reference genome and outputs an .Robj file along with .pdf
network graphs
}
