% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MakeMatrices.R
\name{MakeMatrices}
\alias{MakeMatrices}
\title{Generate matrices for clustering analysis}
\usage{
MakeMatrices(
  isoQuantOutDir = "IsoQuantOutput/",
  groupBy = "Cell",
  ensemblToClear = NULL,
  convertTo10XoutputFormat = TRUE
)
}
\arguments{
\item{isoQuantOutDir}{output directory of IsoQuant
function. Defaults to IsoQuantOutput}

\item{groupBy}{String indicating group for matrix output.
Current options involve "Cell" or "Celltype", and defaults
to "Cell"}

\item{ensemblToClear}{A mapping of gene-names from
ENSEMBL ids to human readable gene names for downstream
analysis. Defaults to mm10 gencode v21 annotation.}

\item{convertTo10XoutputFormat}{OPTIONAL If grouped by cell,
logical inputt indicating whether
or not matrix should be converted into a format similar to 10X
cellranger output for downstream analysis. Defaults to TRUE}
}
\description{
Converts the counts for isoforms
into matrices clustered either by cell-type
or by barcode.
}
\seealso{
\code{\link{IsoQuant}}
}
