% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MapAndAlign.R
\name{MapAndAlign}
\alias{MapAndAlign}
\title{Extract isoform information from aligned reads}
\usage{
MapAndAlign('LRoutput','gencode.vM21.annotation.gtf.gz','chromFa/',16)
}
\arguments{
\item{outputDir}{location within the working directory for files
that will be created during this process}

\item{annoGZ}{annotation.gtf.gz file. Defaults to v21 mm10}

\item{chromFaDir}{directory containing fa.gz files by individual
chromosme. Defaults to mm10}

\item{numThreads}{number of threads for parallel processes. Defaults to 12}
}
\value{
directory containing several bam, gff.gz, and flat files
necessary for downstream analysis
}
\description{
This function follows the STARalign function.
It is a wrapper for several scripts that impose checks and
balances on the mapping quality, consensus nature of splice
sites etc
}
\seealso{
STARalign
}
