% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MapAndFilter.R
\name{MapAndFilter}
\alias{MapAndFilter}
\title{Extract isoform information from aligned reads}
\usage{
MapAndFilter('LRoutput','gencode.vM21.annotation.gtf.gz',16)
}
\arguments{
\item{outputDir}{location within the working directory for files
that will be created during this process}

\item{annoGZ}{annotation.gtf.gz file. Defaults to v21 mm10}

\item{numThreads}{number of threads for parallel processes. Defaults to 12}

\item{seqDir}{directory containing chromosome.fa.gz files}

\item{filterFullLength}{OPTIONAL logical indicating whether reads should be filtered
for having start and end sites falling into annotated cage peaks or polyA
sites. Defaults to FALSE}

\item{cageBed}{OPTIONAL bed.gz file containing annotated CAGE peaks}

\item{polyABed}{OPTIONAL bed.gz file containing annotated polyA sites}

\item{cp_distance}{OPTIONAL distance from annotated CAGE peak or polyA site
for a read to be considered full-length. Defaults to 50}

\item{genomeVersion}{genome and version for GFF output file. Defaults to mm10}
}
\value{
directory containing several bam, gff.gz, and flat files
necessary for downstream analysis
}
\description{
This function follows the alignment function.
It is a wrapper for several scripts that impose checks and
balances on the mapping quality, consensus nature of splice
sites, and annotated start and end sites per transcript.
}
\seealso{
\code{\link{STARalign}} \code{\link{MMalign}}
}
