% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IsoQuant.R, R/utils-pipe.R
\name{\%>\%}
\alias{\%>\%}
\alias{IsoQuant}
\title{Quantifies full-length isoforms in sample(s)}
\usage{
IsoQuant('AllInfo',Iso=TRUE, TSS=FALSE, PolyA=FALSE)

lhs \%>\% rhs
}
\arguments{
\item{Iso}{Logical indicating whether to conduct analysis for full-length
isoforms. Defaults to TRUE}

\item{TSS}{OPTIONAL logical indicating whether to conduct analysis for transcription
start sites (TSS). Defaults to FALSE}

\item{PolyA}{OPTIONAL logical indicating whether to conduct analysis for end sites
(PolyA). Defaults to FALSE}

\item{AllInfo}{Merged file of Gene, Barcode, UMI, Cluster,
and transcript across samples and replicates. Major isoform
is calculated based on the entire input.}
}
\value{
IsoQuantOutput folder containing all useful files for downstream
analysis
}
\description{
Depending on whether you want to do downstream
differential expression analysis of full-length isoforms,
TSS, or PolyA sites, this function annotates and counts up
transcripts expressed in the given set of samples

See \code{magrittr::\link[magrittr:pipe]{\%>\%}} for details.
}
\seealso{
\code{\link{DiffSplicingAnalysis}}
}
\keyword{internal}
