% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiffSplicingAnalysis.R
\name{DiffSplicingAnalysis}
\alias{DiffSplicingAnalysis}
\alias{DiffIsoTest}
\title{Perform differential splicing analysis}
\usage{
DiffSplicingAnalysis('config',10,25,'Exon',is.hier=FALSE)
}
\arguments{
\item{numIsoforms}{Cap on the number of sites to test
per gene. Defaults to 10}

\item{minNumReads}{Parameter to filter out lowly expressed
genes. Defaults to 25}

\item{typeOfTest}{Exon, TSS, PolyA or Isoform test conducted
depending on input. Defaults to Iso}

\item{config}{File in a .tsv format with four columns
specifying the comparisons to be made}
}
\value{
Output folder with results for each line of config
file, i.e. for each set of comparisons

Result file with gene, uncorrected pval, deltaPI, FDR

handy .Robj per comparison to import if need be

file with significantly differentially spliced genes

file with numbers per gene and isoform tested in each group
}
\description{
This function allows you to test alternative
splicing differences between any two groups
based on the full splice-form, exons,
the transcription start site (TSS) or end site (polyA).
}
\seealso{
IsoQuant
}
