% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MapAndAlign.R
\name{MapAndAlign}
\alias{MapAndAlign}
\title{Extract isoform information from aligned reads}
\usage{
MapAndAlign('LRoutput','gencode.vM21.annotation.gtf.gz',16)
}
\arguments{
\item{outputDir}{location within the working directory for files
that will be created during this process}

\item{annoGZ}{annotation.gtf.gz file. Defaults to v21 mm10}

\item{numThreads}{number of threads for parallel processes. Defaults to 12}

\item{filterFullLength}{logical indicating whether reads should be filtered
for having start and end sites falling into annotated cage peaks or polyA
sites. Defaults to FALSE}

\item{cageBed}{bed.gz file containing annotated CAGE peaks}

\item{polyABed}{bed.gz file containing annotated polyA sites}

\item{cp_distance}{distance from annotated CAGE peak or polyA site
for a read to be considered full-length. Defaults to 50}
}
\value{
directory containing several bam, gff.gz, and flat files
necessary for downstream analysis
}
\description{
This function follows the STARalign function.
It is a wrapper for several scripts that impose checks and
balances on the mapping quality, consensus nature of splice
sites etc
}
\seealso{
STARalign
}
