% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigSplitPie.R
\name{sigSplitPie}
\alias{sigSplitPie}
\title{Creates a piechart of number of isoforms needed to cross deltaPI threshold}
\usage{
sigSplitPie('compDir/')
}
\arguments{
\item{compDir}{output directory containing DiffSplicingAnalysis for a
specific comparison}
}
\value{
Directory labelled 'Visualizations' containing a pie chart containing
the breakup of number of isoforms needed to cross deltaPI threshold
of pairwise comparisons containing cell groups provided in input
Legend indicates the number of isoforms needed to cross the deltaPI = 0.1 threshold
}
\description{
A quick way to summarize DIE for a single comparison
}
\seealso{
\code{\link{DiffSplicingAnalysis}}
}
